fix #239 , use python-style templating instead of Jinja in all drivers login_cmd_template
This file assumes Jinja templating for the port parameter and passes a Jinja-style "{{ port }}".

https://github.com/ansible/molecule/blob/main/src/molecule/command/login.py#L105

When it reaches molecule, it is subsituted in  this file and is done by python calling `.format()` on the string.
That causes it to not render correctly and gives users issues running
molecule login.

ref: https://github.com/ansible-community/molecule-plugins/issues/239

Fixed all affected plugins.